requirements.txt
torch
torchvision
scikit-learn
streamlit
ultralytics
matplotlib
numpy
pillow